Add files via upload
Git Description for the Bank  Console Application:

This Git repository contains a console application for Bank. The application simulates a basic banking system with abstract account classes for Savings and Current accounts. Key features of the application include:

1. Abstract `Account` class with properties for ID, Name, and Balance, and methods for Deposit and Withdrawal.
2. Derived classes `SavingsAccount` and `CurrentAccount` with specific withdrawal rules.
3. Notifications for money withdrawal via SMS and email using delegates and events.
4. A `NotificationHandler` class responsible for sending notifications.
5. A `BankOfBaroda` class to organize the bank's functionality, create accounts, and handle notifications.

The application allows for creating and managing bank accounts, making deposits and withdrawals, and receiving notifications when money is withdrawn.

 The code is designed for educational purposes and serves as a basic framework for a simplified bank application.